feat: integrate IcaRAus UNet model, including runner, plotting utilities, and training/testing notebooks.

+++ mmwave_model_integrator/ground_truth_encoders/_gt_encoder.py
@@ -35,7 +35,7 @@ class _GTEncoder:
         """
         self.full_encoding_ready = True
 
-        return gt_data
+        return gt_data.astype(np.uint8)
 
     def reset(self):
         """Implemented by child class to reset encoder history (if applicable)
